Make NTP server configurable
With this change we make the NTP server configurable through the conf.d
infrastructure. That is, a file /etc/conf.d/ntpdate setting the
NTP_SERVER variable can be used to set the NTP server to use.
Furthermore, if no such file exists we fall back to the new default of
time.nist.gov.